+++ src/openai/runConcurrent.ts
@@ -1,4 +1,3 @@
-import { Chat } from "openai/resources";
 import { printChatMessages, printMarkdown } from "../helpers/print";
 import { ChatMessage } from "./messages";
 import * as _ from "remeda";